Revert "refactor(codegen): extract helpers and annotate all branches for maintainability"

+++ wrappers/java/ratchet/src/main/java/com/virgilsecurity/crypto/ratchet/RatchetMessage.java
@@ -1,5 +1,5 @@
 /*
-* Copyright (C) 2015-2026 Virgil Security, Inc.
+* Copyright (C) 2015-2022 Virgil Security, Inc.
 *
 * All rights reserved.
 *
@@ -7,17 +7,17 @@
 * modification, are permitted provided that the following conditions are
 * met:
 *
-*     (1) Redistributions of source code must retain the above copyright
-*     notice, this list of conditions and the following disclaimer.
+* (1) Redistributions of source code must retain the above copyright
+* notice, this list of conditions and the following disclaimer.
 *
-*     (2) Redistributions in binary form must reproduce the above copyright
-*     notice, this list of conditions and the following disclaimer in
-*     the documentation and/or other materials provided with the
-*     distribution.
+* (2) Redistributions in binary form must reproduce the above copyright
+* notice, this list of conditions and the following disclaimer in
+* the documentation and/or other materials provided with the
+* distribution.
 *
-*     (3) Neither the name of the copyright holder nor the names of its
-*     contributors may be used to endorse or promote products derived from
-*     this software without specific prior written permission.
+* (3) Neither the name of the copyright holder nor the names of its
+* contributors may be used to endorse or promote products derived from
+* this software without specific prior written permission.
 *
 * THIS SOFTWARE IS PROVIDED BY THE AUTHOR ''AS IS'' AND ANY EXPRESS OR
 * IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
@@ -36,24 +36,36 @@
 
 package com.virgilsecurity.crypto.ratchet;
 
+import com.virgilsecurity.crypto.foundation.*;
+
+/*
+* Class represents ratchet message
+*/
 public class RatchetMessage implements AutoCloseable {
 
     public long cCtx;
 
+    /* Create underlying C context. */
     public RatchetMessage() {
         super();
         this.cCtx = RatchetJNI.INSTANCE.ratchetMessage_new();
     }
 
+    /* Wrap underlying C context. */
     RatchetMessage(RatchetContextHolder contextHolder) {
         this.cCtx = contextHolder.cCtx;
     }
 
+    /*
+    * Acquire C context.
+    * Note. This method is used in generated code only, and SHOULD NOT be used in another way.
+    */
     public static RatchetMessage getInstance(long cCtx) {
         RatchetContextHolder ctxHolder = new RatchetContextHolder(cCtx);
         return new RatchetMessage(ctxHolder);
     }
 
+    /* Clear resources. */
     private void clearResources() {
         long ctx = this.cCtx;
         if (this.cCtx > 0) {
@@ -62,48 +74,77 @@ public class RatchetMessage implements AutoCloseable {
         }
     }
 
+    /* Close resource. */
     public void close() {
         clearResources();
     }
 
+    /* Finalize resource. */
     protected void finalize() throws Throwable {
         clearResources();
     }
 
+    /*
+    * Returns message type.
+    */
     public MsgType getType() {
         return RatchetJNI.INSTANCE.ratchetMessage_getType(this.cCtx);
     }
 
-    public int getCounter() {
+    /*
+    * Returns message counter in current asymmetric ratchet round.
+    */
+    public long getCounter() {
         return RatchetJNI.INSTANCE.ratchetMessage_getCounter(this.cCtx);
     }
 
+    /*
+    * Returns long-term public key, if message is prekey message.
+    */
     public byte[] getSenderIdentityKeyId() {
         return RatchetJNI.INSTANCE.ratchetMessage_getSenderIdentityKeyId(this.cCtx);
     }
 
+    /*
+    * Returns long-term public key, if message is prekey message.
+    */
     public byte[] getReceiverIdentityKeyId() {
         return RatchetJNI.INSTANCE.ratchetMessage_getReceiverIdentityKeyId(this.cCtx);
     }
 
+    /*
+    * Returns long-term public key, if message is prekey message.
+    */
     public byte[] getReceiverLongTermKeyId() {
         return RatchetJNI.INSTANCE.ratchetMessage_getReceiverLongTermKeyId(this.cCtx);
     }
 
+    /*
+    * Returns one-time public key, if message is prekey message and if one-time key is present, empty result otherwise.
+    */
     public byte[] getReceiverOneTimeKeyId() {
         return RatchetJNI.INSTANCE.ratchetMessage_getReceiverOneTimeKeyId(this.cCtx);
     }
 
+    /*
+    * Buffer len to serialize this class.
+    */
     public int serializeLen() {
         return RatchetJNI.INSTANCE.ratchetMessage_serializeLen(this.cCtx);
     }
 
+    /*
+    * Serializes instance.
+    */
     public byte[] serialize() {
         return RatchetJNI.INSTANCE.ratchetMessage_serialize(this.cCtx);
     }
 
+    /*
+    * Deserializes instance.
+    */
     public static RatchetMessage deserialize(byte[] input) throws RatchetException {
         return RatchetJNI.INSTANCE.ratchetMessage_deserialize(input);
     }
-
 }
+
